Purge dead code and disposable diagnostics; zero compiler warnings
Cleanup pass done alongside the ALU sign-extension audit, on the theory
that dead/dirty code makes this kind of systemic-bug hunting harder.

- Removed 7 write-only fields with zero reads anywhere: Cdvd._streamRemaining,
  EmotionEngine._nullifyDelayIfNotTaken, Gif._fifoWords, Iop._branchPending,
  MidwayBootAssist._esrbDone, Ps2System._commercialWorkerKicked, Spu2._irqPending.
- Removed Program.cs's 19 disposable probe-* commands (hardcoded to a specific
  dev machine's paths and MK Shaolin Monks' addresses, already documented as
  "not a maintained, stable tool surface"). Kept probe-frame and the
  extract-file/elf-sections/iop-disasm/iop-find-word commands, all of which
  are actively documented, general-purpose tools. Program.cs: 1865 -> 869 lines.
- Fixed all remaining compiler warnings (12 -> 0 across Core/Tests/Desktop):
  an unused local in HostGamepad.cs, two CS0675 sign-extension warnings
  (GsRegisters.PackScissor, a SmokeTests helper) fixed by widening through
  uint first, a CS8602 in GameDisplayWindow.axaml.cs, and the CS8600/CS8602
  chain in Program.cs's elf-info command caused by SystemCnf.Boot2 being
  nullable.

Full smoke suite green; verified every previously-documented stable CLI
command still exists by diffing command names against the last commit.

+++ Tests/SmokeTests.cs
@@ -546,8 +546,8 @@ public static class SmokeTests
 
     private static void WriteAdXyz(SystemMemory mem, ref uint addr, int x, int y, uint z)
     {
-        ulong xyz = (ulong)((x << 4) & 0xFFFF)
-                  | ((ulong)((y << 4) & 0xFFFF) << 16)
+        ulong xyz = (ulong)(uint)((x << 4) & 0xFFFF)
+                  | ((ulong)(uint)((y << 4) & 0xFFFF) << 16)
                   | ((ulong)(z & 0xFFFFFF) << 32);
         WriteAd(mem, ref addr, 0x04, xyz);
     }
